% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.MultiTaskELR.R
\name{predict.MultiTaskELR}
\alias{predict.MultiTaskELR}
\title{predict.MultiTaskELR}
\usage{
\method{predict}{MultiTaskELR}(object, newdata, class.type = "class", ...)
}
\arguments{
\item{object}{Trained MultiTaskELR model.}

\item{newdata}{A a named list of matrices
for each task all of the same dimension and type of predictors.
Names should be exactly the same as the names of the response variables
used for training.}

\item{class.type}{type of prediction for classification task:
"class" for class memberships or "prob" for class probabilities.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
predictions: a list with predictions
\item{class.pred}{a matrix whose columns are the predicted class
memberships or probabilities for classification tasks}
\item{reg.pred}{a matrix whose columns are the predictions for
regression tasks}
}
\description{
Make predictions for MTL with ELR and/or ELM for new data
}
\author{
Che Ngufor <Ngufor.Che@mayo.edu>
}
\references{
[1] Che Ngufor, Sudhindra Upadhyaya, Dennis Murphree, Nageswar R. Madde,
Daryl J. Kor, and Jyotishman Pathak. "A Heterogeneous Multi-task
Learning for Predicting RBC Transfusion and Perioperative Outcomes".
To appear in the  15th Conference on Artificial Intelligence in Medicine,
AIME 2015, Pavia, Italy, June 17-20, 2015. Proceedings.
}

