% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils}
\alias{Kernel.HW}
\alias{SynData.MultiTaskELR}
\alias{Utils}
\title{Utility functions}
\usage{
Kernel.HW(form, X, ken, p, H.W = NULL)

SynData.MultiTaskELR(n = 100, d = 10, n.class = 3, n.reg = 2,
  class.prior = 0.5)
}
\arguments{
\item{form}{formula}

\item{X}{data matrix}

\item{ken}{kernel type}

\item{p}{dimension of random feature space}

\item{H.W}{random weights as computed by \code{\link[ELR]{Weights.ELR}}}

\item{n}{sample size}

\item{d}{dimension}

\item{n.class}{number of classifications tasks}

\item{n.reg}{number of regression tasks}

\item{class.prior}{proportion of positive examples}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Randomize kernel matrix from \code{Kernel.HW } and list of
}
\description{
Generate sequence of artificial binary classification data
 from two multivariate normal distribution  for multitask
 learning. All datasets have the
 same number of examples n and the same dimension  d
}
\author{
Che Ngufor <Ngufor.Che@mayo.edu>
}
\references{
Che Ngufor, Sudhindra Upadhyaya, Dennis Murphree, Nageswar R. Madde,
Daryl J. Kor, and Jyotishman Pathak. "A Heterogeneous Multi-task
Learning for Predicting RBC Transfusion and Perioperative Outcomes".
To appear in the  15th Conference on Artificial Intelligence in Medicine,
AIME 2015, Pavia, Italy, June 17-20, 2015. Proceedings.
}

